Added a helper method to round prices and volumes to within the correct size. Also changes smart trade properties over from long to ints.

+++ src/XCommasSharp/XCommasSharp/Models/SmartTrade.cs
@@ -10,10 +10,10 @@ namespace XCommasSharp.Models
     public class SmartTrade
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("version")]
-        public long Version { get; set; }
+        public int Version { get; set; }
 
         [JsonProperty("account")]
         public Account Account { get; set; }
